10/requirements.txt
amqp==5.0.6
anyjson==0.3.3
asgiref==3.3.4
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
billiard==3.6.4.0
blinker==1.4
cached-property==1.5.2
celery==5.1.0
certifi==2018.1.18
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cloud-init==20.4
colorama==0.3.7
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.1.4
distro-info===0.18ubuntu0.18.04.1
Django==3.2.4
django-appconf==1.0.4
django-celery-beat==2.2.0
django-celery-email==3.0.0
django-celery-results==2.0.1
django-timezone-field==4.1.2
djangorestframework==3.12.4
ephem==3.7.6.0
httplib2==0.9.2
hyperlink==17.3.1
idna==2.6
importlib-metadata==4.4.0
incremental==16.10.1
Jinja2==2.10
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
keyring==10.6.0
keyrings.alt==3.0
kombu==5.1.0
language-selector==0.1
MarkupSafe==1.0
netifaces==0.10.4
oauthlib==2.0.6
PAM==0.4.2
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycrypto==2.6.1
pygobject==3.26.1
PyJWT==1.5.3
pyOpenSSL==17.5.0
pyparsing==2.2.0
pyserial==3.4
python-apt==1.6.5+ubuntu0.5
python-crontab==2.5.1
python-dateutil==2.8.1
python-debian==0.1.32
python-memcached==1.53
pytz==2021.1
pyxdg==0.25
PyYAML==3.12
redis==3.5.3
requests==2.18.4
requests-unixsocket==0.1.5
SecretStorage==2.3.1
service-identity==16.0.0
six==1.16.0
sqlparse==0.4.1
ssh-import-id==5.7
systemd-python==234
tornado==6.1
Twisted==17.9.0
typing-extensions==3.10.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.22
vine==5.0.0
virtualenv==15.1.0
wcwidth==0.2.5
zipp==3.4.1
zope.interface==4.3.2
